refactor: Add "Load More" button to Products page
The code changes add a "Load More" button to the Products page in the user section. This button allows users to load more products dynamically without refreshing the page. The button is implemented using the MUI Button component and includes a refresh icon. This enhancement improves the user experience by providing a convenient way to view more products.