refactor: Update Motion components usage in user pages
The code changes in multiple user pages update the usage of Motion components. The Stack components are replaced with StackM components, and the createMotionProps function is used to pass motion props to the StackM components. This refactor improves the animation and motion effects in the user pages, enhancing the overall user experience.